Reduce generated content
