+++ package.json
@@ -23,6 +23,10 @@
     "prepare": "husky install",
     "example": "node -r esbuild-register ./example/server.ts"
   },
+  "dependencies": {
+    "lodash-es": "^4.17.21",
+    "type-fest": "^2.18.0"
+  },
   "peerDependencies": {
     "mjml": "^4.13.0",
     "mjml-core": "^4.13.0",
@@ -33,10 +37,11 @@
     "@changesets/cli": "^2.24.2",
     "@fransvilhelm/changeset-changelog": "^1.1.1",
     "@trivago/prettier-plugin-sort-imports": "^3.3.0",
-    "@types/lodash": "^4.14.182",
+    "@types/lodash-es": "^4.17.6",
     "@types/mjml": "^4.7.0",
     "@types/mjml-core": "^4.7.1",
     "@types/node": "^18.6.5",
+    "@types/prettier": "^2.7.0",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "cheerio": "^1.0.0-rc.12",
@@ -46,7 +51,6 @@
     "eslint-config-react-app": "^7.0.1",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3",
-    "lodash": "^4.17.21",
     "micro": "^9.4.1",
     "mjml": "^4.13.0",
     "mjml-core": "^4.13.0",
