+++ src/components.tsx
@@ -1,51 +1,5 @@
-import _ from 'lodash';
-import React from 'react';
-
-type MjUnitBase<Suffix extends string> = `${number}${Suffix}` | '0';
-export type MjUnit<Suffix extends string> =
-  | `${MjUnitBase<Suffix>}`
-  | `${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>}`
-  | `${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>}`
-  | `${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>} ${MjUnitBase<Suffix>}`;
-
-export type MjmlBaseProps = {
-  /** class name, added to the root HTML element created */
-  className?: string;
-  /** class name, added to the root HTML element created */
-  cssClass?: string;
-  children?: React.ReactNode;
-};
-
-export type MjmlComponent<Props> = React.FC<Props & MjmlBaseProps>;
-
-type KebabCase<T extends string, A extends string = ''> = T extends `${infer F}${infer R}`
-  ? KebabCase<R, `${A}${F extends Lowercase<F> ? '' : '-'}${Lowercase<F>}`>
-  : A;
-
-type KebabCaseKeys<T> = { [K in keyof T as K extends string ? KebabCase<K> : K]: T[K] };
-
-function handleMjmlProps<T extends Record<string, unknown>>(props: T): KebabCaseKeys<T> {
-  let converted: Record<string, unknown> = {};
-  for (let [key, value] of Object.entries(props)) {
-    converted[_.kebabCase(key)] = value;
-  }
-  return converted as KebabCaseKeys<T>;
-}
-
-function createComponent<Props extends Record<string, unknown>>(Name: string): MjmlComponent<Props> {
-  const Component: MjmlComponent<Props> = ({ className, cssClass, children, ...rest }) => {
-    let props = handleMjmlProps(rest);
-    return (
-      // @ts-expect-error
-      <Name {...props} css-class={cssClass ?? className}>
-        {children}
-      </Name>
-    );
-  };
-  Component.displayName = Name;
-
-  return Component;
-}
+import { createComponent } from './create-component';
+import { MjUnit } from './types';
 
 export type MjBodyProps = {
   /** email's width */
