+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './extensions';
 export * from './render';
+export * from './types';
